Create Bobbys hobbies
I've created a database of people and hobbies, and each row in hobbies is related to a row in persons via the person_id column. Then i Select two tables with a join so that you can see each person's name next to their hobby. Lastly, i added an additional query that shows only the name and hobbies of 'Bobby McBobbyFace', using JOIN combined with WHERE.