Create Customers orders list
i've created a database for customers and their orders. Not all of the customers have made orders, however. Come up with a query that lists the name and email of every customer followed by the item and price of orders they've made. I used a LEFT OUTER JOIN so that a customer is listed even if they've made no orders. Then i created another query that will result in one row per each customer, with their name, email, and total amount of money they've spent on orders.