Create Sequels in SQL
Here i've created a table with all the 'Harry Potter' movies, with a sequel_id column that matches the id of the sequel for each movie. I put down a SELECT that will show the title of each movie next to its sequel's title (or NULL if it doesn't have a sequel).