Create FriendBook
In this project, i've created a database for a friend networking site, with a table storing data on each person, a table on each person's hobbies, and a table of friend connections between the people. I used a JOIN to display a table showing people's names with their hobbies. Then i used another SELECT with a JOIN to show the names of each pair of friends, based on the data in the friends table.